requirements.txt
# Web framework
flask==2.3.3

# Data & computation
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# Torch & model
torch==2.1.2

# Image handling
Pillow==10.3.0

# CLIP from OpenAI GitHub (not pip clip!)
ftfy==6.1.1
regex==2023.12.25
tqdm==4.66.4
git+https://github.com/openai/CLIP.git

# Optional (faster CSV loading if used)
# pyarrow==15.0.2

